requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.1
matplotlib==3.8.0
pandas==2.1.2
requests==2.31.0
pyyaml==6.0.1
networkx==3.2.1
scikit-learn==1.3.2
scipy==1.11.3
conda install pytorch::pytorch torchvision torchaudio -c pytorch
torch-geometric==2.4.0